# docs/swagger.yaml
definitions:
  application_dtos.CreateApplicationDTO:
    properties:
      audience_id:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - audience_id
    - name
    type: object
  entities.Application:
    properties:
      audience:
        $ref: '#/definitions/entities.ResourceServer'
      client_id:
        type: string
      client_secret:
        items:
          $ref: '#/definitions/entities.ClientSecret'
        type: array
      created_at:
        type: integer
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: integer
    type: object
  entities.ClientSecret:
    properties:
      created_at:
        type: integer
      expire_at:
        type: integer
      name:
        type: string
      value:
        type: string
    type: object
  entities.ResourceServer:
    properties:
      created_at:
        type: integer
      description:
        type: string
      display_name:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: integer
    type: object
  entities.User:
    properties:
      created_at:
        type: integer
      email:
        type: string
      id:
        type: string
      password:
        type: string
      updated_at:
        type: integer
    type: object
  resource_server_dtos.CreateResourceServerDTO:
    properties:
      description:
        type: string
      display_name:
        type: string
    type: object
  resource_server_dtos.UpdateResourceServerDTO:
    properties:
      description:
        type: string
      display_name:
        type: string
    type: object
  token_dtos.AccessTokenResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      token_type:
        type: string
    type: object
  user_dtos.CreateUserDTO:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - confirm_password
    - email
    - password
    type: object
info:
  contact: {}
paths:
  /applications/:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all applications
      parameters:
      - default: 10
        description: Number of applications to return
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/entities.Application'
              type: array
            type: array
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all applications with pagination
      tags:
      - Applications
    post:
      consumes:
      - application/json
      description: Create a new application with the provided name, description, and
        audience ID
      parameters:
      - description: Application creation data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/application_dtos.CreateApplicationDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Application'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create a new application
      tags:
      - Applications
  /applications/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve an application by its ID
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Application'
        "404":
          description: Not Found
          schema: {}
      summary: Get application by ID
      tags:
      - Applications
    put:
      consumes:
      - application/json
      description: Update an existing application's name and description
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Application update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/application_dtos.CreateApplicationDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Application'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update an existing application
      tags:
      - Applications
  /resource-servers/:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of resource servers
      parameters:
      - default: 10
        description: Number of resource servers to return
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.ResourceServer'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all resource servers with pagination
      tags:
      - ResourceServers
    post:
      consumes:
      - application/json
      description: Create a new resource server with the provided display name and
        description
      parameters:
      - description: Resource server creation data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resource_server_dtos.CreateResourceServerDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.ResourceServer'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create a new resource server
      tags:
      - ResourceServers
  /resource-servers/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a resource server by its ID
      parameters:
      - description: Resource Server ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ResourceServer'
        "404":
          description: Not Found
          schema: {}
      summary: Get resource server by ID
      tags:
      - ResourceServers
    put:
      consumes:
      - application/json
      description: Update an existing resource server's display name and description
      parameters:
      - description: Resource Server ID
        in: path
        name: id
        required: true
        type: string
      - description: Resource Server update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resource_server_dtos.UpdateResourceServerDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ResourceServer'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update an existing resource server
      tags:
      - ResourceServers
  /token/:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Dispatch tokens based on the provided grant type
      parameters:
      - description: Grant type
        in: formData
        name: grant_type
        required: true
        type: string
      - description: Username for password grant
        in: formData
        name: username
        type: string
      - description: Password for password grant
        in: formData
        name: password
        type: string
      - description: Client ID for client credentials grant
        in: formData
        name: client_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/token_dtos.AccessTokenResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Token dispatch endpoint
      tags:
      - Tokens
  /users/:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided email and password
      parameters:
      - description: User creation data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user_dtos.CreateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create a new user
      tags:
      - Users
swagger: "2.0"